refactor(ndt7): move logging in its own package
